+++ packages/browser-wallet/src/popup/pages/Account/Tokens/ManageTokens/ManageTokens.scss
