+++ packages/browser-wallet/src/popup/pages/Account/Tokens/ManageTokens/state.ts
