+++ packages/browser-wallet/src/popup/pages/Account/index.scss
@@ -7,6 +7,6 @@
 @import './TransactionLog/TransactionLog';
 @import './SendCcd/SendCcd';
 @import './Tokens/Tokens';
-@import './ManageTokens/ManageTokens';
+@import './Tokens/ManageTokens/ManageTokens';
 @import './Earn/Earn';
 @import './ConfirmTransfer/ConfirmTransfer';
